neovim(readline): Plugin for emacs/readline style keybinds
This plugin adds keybinds for insert mode and command mode.
The keybinds allow to use emacs style keybinds for editing.
Which I find convenient, specially for command mode where it is not possible to execute normal commands
But also for Insert mode, since I use these keybinds everywhere 'readline (3)' is used in Vim

For the moment there are mainly keybinds to move the cursor around, such as:
    * forward-char and backward-char
    * forward-word and backward-word
    * end-of-line and beggining-of-line
As well as keybinds to kill a region and the yank command:
    * Kill-word and backward-kill-word
    * Kill-line and backward-kill-line
    * Delete-char and backward-delete-char
I implemented two version for most of these functions.
One for Insert mode and the other for Command mode
The reason is, for Insert mode these functions can be implemented with normal commands for the most part
and that is simply achieved by typing <C-o> in insert mode.
Whereas for command mode, everything had to be implemented using builtin function to manipulate the command line
And '<C-\>e' to replace the command line inside command mode mappings