neovim(map): New insert mode mapping
CleverTab is taken from the help files.
Allows to TAB like normal at the beginning of a line, or start completion when in front of a word
Deleted the file .config/nvim/plugin/map.lua,
which had a previous TAB remap that was similar but less useful

The BufReadPost autocmd is also taken from the help file
But the pattern is different, it is adapted to the pattern I use for backup file

Mapped Ctrl-n and Ctrl-p to Down/Up respectively in insert mode
to make vertical movement and small adjustements easier.

For convenience, remap <S-j> and <S-k> in visual mode to j and k respectively
Reason is, I often accidentaly press the key with Shift activated
after linewise visual mode instead of simply pressing j/k to move down/up
I dont use the default mapping in visual mode anyway.
Also remapped <S-j> to simply j in normal mode, as I rarely use that key.

For command mode, remapped <C-x> to <C-a>, since <C-a> was remapped already
and it could be useful.