nvim(gui): Update font used
The fonts specified were not up to date with the one present in this repo
So this commit merely fixes the font setting used in Neovim, to be synced with the fonts
located in "~/.local/share/fonts"

In general, I use the normal NerdFont patched font.
I dont monospaced since my terminal handles it well.
And the Propo does not make a big difference.