(nvim) Remove fancy nerd font separator from lualine
They look bad with a transparent background because the "arrow" wont be transparent.
It is supposed to meld with the background color, that's why it has the same color.
But since its a foreground character, no transparency is applied on it. So it will stop blending.

For example if we set transparency to 100% (no background), the background will be invisible.
But we still have some separators showing in the middle of the screen.