(lsp) Add language server support for java and lua
Add plugin for nvim to start and configure the language servers
and script on the global PATH to start the language server with the right
command line option.

For java, also update the environment variables. It now it fix and points to a
symlink that never moves. Its value changes to point to the current version
of the jdk used. Also update 'mvn' environment variable and script