(lsp) Add 'root_dir' for LuaLS
The language server was sometimes not able to run. Because it did not always have a root dir
It is now rooted in my neovim config directory. For that I need to provide a root marker.
I added 'init.vim' and 'init.lua' as markers and created an empty 'init.vim' file.