(nvim) Smarter command to edit config files
Improve command to open neovim configuration folder.
Whether to edit or just find a specific file/information. This is something I do a lot
so improving that flow is important.

This is done by opening my config in a specific tab/window and keeping track of the id
And that window are local specific options and settings to make it easier to open the files I want.
I ignore the 'pack' folder, since I dont want to edit an imported package.
Set 'pwd' for the tab in the configuration folder, so that ':find' have very few files to look for.

Also set 'grepprg' to 'git grep'. And I can now easily find a specific line of configuration after
going to the config tab