(nvim) Clean up some plugins
Mainly removing bang from functions. It is only useful when sourcing the file
But with ':runtime' I can modify the file and re-read it.
The bang is ugly

And renamed the 'Get_char()' function. It was the first function of vimscript in my config
Im starting to converge on a specific style for the vimscript I write. Just making it adhere to it