Merge old 'neovim-packages' repository into 'neovim-config'
For unifying all my neovim related configuration, I changed my repository setup
All of the neovim config besides the 'pack/' folder is taken from my main dotfiles repository at
'https://github.com/ryanbeerbarron/dotfiles.git' and with 'git filter-repo' captured the subtree
I was interested in and kept the history.

The 'pack/' folder comes from the now archived repository
'https://github.com/ryanbeerbarron/neovim-packages' with a few modification that were done before archiving namely:
1. Moving the root of the repository from 'git-modules/**' to 'nvim/pack/git-modules/**'
2. Splitting packages into different directories based on different criterias
3. Removing lightline.vim which was unused.
4. Simplified the logical name of the submodules

The first step was done with 'git filter-repo'. While the commits are recreated, it keeps the history intact.
And the last step was done because the logical name were based on the previous path with does not make sense
after it was changed. This was done by doing 'git deinit --all' and the manually modifying the '.gitmodules' file.

After that, to merge the two repository, I added the 'neovim-packacges' repository as a remote to the new one
and merge both branch master branch with the '--allow-unrelated-histories' flag on.