(projects) Always create file/directory if not present
While updating to nvim v0.10, I deleted some old 'nvimrc' file I was using.

The plugin was reporting errors on startup if any of the file were missing.
The files were only created the first time a project was opened. This now will
always check that the files are present and recreate them if needed