(workspace) Add option to not load any workspace
When writing git commit messages, I dont want to open a new instance of neovim with the same workspace.
I get errors messages when trying to reopen all the buffers, since they are already open.

Using a neovim server per workspace, via the '--listen' command option and the 'remote' features does not work for my use case
since git launch the '$EDITOR' for commit editing and is waiting on the child process to finish.

I added a default `colorscheme` command for those neovim instance that don't run in a workspace with a special shadafile