Rewrite 'workspace' plugin completely
It is now a lua module and is the first module executed in 'init.lua'
The module only works if the global variable 'long_session' is set which is done
by a 'nvim' wrapper 'nvim_session' in my main dotfiles repository.

The idea is to have two type of 'nvim' executable:
    * One for long running session with all the LSP needed and specific configuration
      for a particular project. Goal is to be closer to a fully fledge IDE.
    * A lightweight version of nvim for quick edits, closer to vanilla nvim

The lightweight version is still invoked with simply 'nvim' and the IDE version with 'nvim_session'

For the main part, the plugin does the same as before.
It sets a specific 'shadafile' and `nvimrc` (see ':h mkvimrc')
but now also sets a specific 'directory' and 'undodir'.
The workspace specific config is now added to 'rtp' and the main rtp directories are created.