Add new functions for setting 'foldtext'
When folding a function, I often want to fold the comments that are above
but this result in confusing 'foldtext' output. Where the function name is not shown
and only the comments.
The two function allow me to ignore the comments and still show the function name in the 'foldtext'
output.
This is usually set with a modeline