Add 'Config' and 'ConfigGrep' command
Commands to quickly search my own neovim configuration.
'Config' will simply open the file passed as argument and
'ConfigGrep' will search for a pattern in my configuration files, similar to
the builtin 'helpgrep'.

Most of the heavy work is made by the custom completion to quickly find
which file I want to open.
It does so by globbing all the relevant files ( and sets a temporary 'wildignore' )
and uses the text typed on the command line to match.