(options) Remove some default '&formatoptions' flags
The 'l', 'c', and 't' which are removed caused automatic linebreaks and auto-wrap
I want to have control on how text looks and is inserted so I prefer to write without them.

These flags are set by nvim default configuration, in '$VIMRUNTIME'.
The only way to override them is with ftplugin in the 'after/' directory.