Removing LSP submodules
The java Lsp packages have not been used in a while and a complicated to configure.
And with Neovim 1.0 approaching, a lot of new features and settings are there out of the box.

To prepare myself to learn and relearn Neovim for 1.0, I prefer removing parts of my configuration
that are too complicated to maintain in its current state.